Re-tooling kd_tree.
Eliminated space-filling kD-trees.  All trees now store a
space-filling bound.  If bounds that are tight on the enclosed objects
are required, kd_tree exports bounds_of_objects.  We ensure that there
is always at least one object in every cell by splitting at the median
(or just to the left or right of the median if the median would put
all objects on one side or the other).