Added admixture test to MCMC.
Must be careful in admixture MCMC to not propose jumps without prior
support; if you do so, then the distribution can just wander off.
Lack of prior support in one component of the model does *not* give
zero posterior, but instead cuts the posterior in half (if the volumes
of the two models are about equal).  So, one must make sure in the
jump proposal that one does not go outside the region of prior support
of each model.